fix for remote timestamps in the past longer than 3 seconds ago

+++ src/icmp_server.cpp
@@ -72,7 +72,7 @@ void run_in(shm_message_queue *const shm, const std::string & out_name)
 
 				uint64_t now_local = get_us();
 
-				if (labs(tv.tv_sec - now_local / 1'000'000) < 3) {
+				if (labs(int64_t(tv.tv_sec - now_local / 1'000'000)) < 3) {
 					int64_t age = now_local - (tv.tv_sec * 1'000'000 + tv.tv_usec);
 					age_str += " (sent " + std::to_string(age) + " µs ago)";
 				}
